backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-dotenv>=1.1.0
python-multipart>=0.0.24
pdfplumber>=0.11.8
python-docx>=1.1.1
supabase>=2.10.0
pinecone
llama-index
llama-index-vector-stores-pinecone
llama-index-llms-groq
PyJWT>=2.10.0
httpx>=0.28.0
pydantic-settings
cryptography
pytest>=8.0.0
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0
slowapi>=0.1.9
